requirements.txt
pytest==5.3.5
Appium-Python-Client==2.0.0a0
allure-pytest
python-decouple==3.4